+++ src/test/testData/navigation/attribute.php
@@ -4,3 +4,7 @@ namespace App\Attribute;
 
 #[\Attribute]
 class Handler {}
+
+// Атрибут-маркер класса-сообщения (правило фильтра по атрибуту, не по типу).
+#[\Attribute]
+class AsMessage {}